refactored the parameters of a few commands, to be more concrete, added attributes to channel group overview, added command to see the current configuration of a channel, added comand to remove channel group, refactored commands from administration to own module, made the range for the random xp configurable in the database, adapted update level command to also work for single users, added command to configure the role of a level, fixed any reaction triggering the removal in the profanity report, removed the id column from the role table, added xp role column to the role table

+++ src/OnePlusBot/Data/Models/ExperienceRoles.cs
@@ -14,7 +14,7 @@ namespace OnePlusBot.Data.Models
         public uint Id { get; set; }
 
         [Column("role_id")]
-        public uint ExperienceRoleId { get; set; }
+        public ulong ExperienceRoleId { get; set; }
         
         [Column("level")]
         public uint Level { get; set; }
